task 8 completed - add new comment
